+++ src/librustdoc/test.rs
@@ -67,7 +67,7 @@ pub fn run(options: Options) -> i32 {
     cfgs.push("doctest".to_owned());
     let config = interface::Config {
         opts: sessopts,
-        crate_cfg: config::parse_cfgspecs(cfgs),
+        crate_cfg: interface::parse_cfgspecs(cfgs),
         input,
         input_path: None,
         output_file: None,
